Fly Setup in progress and added mail service

+++ server/redis.js
@@ -10,6 +10,7 @@ function initRedisConnection() {
   return new Promise((resolve, reject) => {
     let promise;
     // eslint-disable-next-line no-undef
+    console.log('process.env', process.env.REDIS_URL)
     redisClient = createClient({
       // eslint-disable-next-line no-undef
       url: process.env.REDIS_URL
@@ -71,7 +72,12 @@ export function getAllCachedDataByKey(key) {
 
 export function deleteAllCaches() {
   console.log('Deleting caches');
-  return redisClient.flushDb();
+  if (redisClient) {
+    return redisClient.flushDb();
+  } else {
+    return null;
+  }
+
 }
 
 export default initRedisConnection;
